sbi: Define tvm_initiate_fence()
There are some memory management operations in a TVM's lifecycle -- page
promotion/demotion and interrupt file unassignment -- that require the
host to invalidate pages in a TVM. We can use a similar procedure as the
one used to flush host translations when converting pages to flush TVM
translations.

Define a new tvm_initiate_fence() call in the TEE-Host API, which is
analogous to the tsm_initiate_fence() call in the conversion fence
procedure. Unlike the conversion fence procedure, the host need not enter
the TSM on each physical CPU in order to complete the flush. Instead, the
flush is done by automatically by the TSM the next time a vCPU in the
target TVM is run. Thus the TVM fence can be considered complete once
every vCPU that was running prior to the call to tvm_initiate_fence()
has taken a trap into the TSM. In pseudo-code, such a sequence might
look like:

atomically {
    tvm_initiate_fence(vm->id);
    for_each_vcpu(vcpu, vm) {
        if (vcpu->is_running) {
	    send_ipi(vcpu->physical_cpu);
	}
    }
}

This sequence will be used in follow-on patches to support unbinding of
IMSIC interrupt files from a TVM vCPU.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>